fix: address code review feedback for conversation history persistence
- Add composite index on memory_facts(user_id, last_seen DESC) for query performance
- Wrap upsertFact in atomic transaction to prevent race conditions
- Optimize fact persistence to skip unnecessary upserts
- Replace busy-wait with Bun.sleep() in tests
- Extract conversation management to conversation.ts to fix max-lines lint warning
